Make cw721-base minter updatable via three-step ownership transfer.
This makes a breaking change to cw721-base in that it makes the
`minter` field on `cw721_base::MinterResponse` optional.

Before:

```rust
pub struct MinterResponse {
    pub minter: String,
}
```

After:

```rust
pub struct MinterResponse {
    pub minter: Option<String>,
}
```

It also makes a breaking change to the `ContractError` enum in that it
removes the `Unauthorized` variant in favor of cw-ownable's more
descriptive `OwnershipError`.

It then extends the `cw721_base`, but not the `cw721` standard's,
`ExecuteMsg` with
[`cw-ownable`](https://crates.io/crates/cw-ownable)'s
`UpdateOwnership` enum variant, and extends the `QueryMsg` with
`cw-ownable`'s `Ownership {}` query. Using these messages, the current
minter on the contract may nominate a new minter, who may accept or
reject this nomination.